move create cmd tests to own module
